[ZEDAPP-31]Added Order service

+++ src/main/java/com/zedapp/repository/GenericRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @NoRepositoryBean
-public interface GenericRepository<T> extends CrudRepository<T, Integer> {
+public interface GenericRepository<T> extends CrudRepository<T, Long> {
     @Override
     List<T> findAll();
